pyproject.toml
[tool.poetry]
name = "loan-risk-predictor"
version = "0.1.0"
description = "Loan Risk Prediction System with Multiple Models"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.0"
pandas = "^2.2.0"
scikit-learn = "^1.4.0"
torch = "^2.2.0"
lightgbm = "^4.3.0"
customtkinter = "^5.2.2"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pyyaml = "^6.0.1"
tqdm = "^4.67.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
